requirements.txt
streamlit==1.28.1
pymongo==4.5.0
dnspython==2.4.2
firebase-admin==6.2.0
python-dotenv==1.0.0
bcrypt==4.0.1
pyrebase4==4.7.1
plotly.express==0.4.1
databases==0.6.1